+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Argon2id support
 - System wide values option
 - Allow email login option
+- UID user table column
 
 ## [4.4.1] - 2020-02-02
 ### Fixed